+++ run_bot.py
@@ -9,6 +9,6 @@ django.setup()
 
 
 if __name__ == "__main__":
-    from bot.handlers.base import run_bot
+    from bot.bot import run_bot
 
     run_bot()
